API update: Fixed dependency resolution Fixed return types for template types

+++ api/CExoLinkedList.cpp
@@ -1,19 +1,5 @@
 #include "CExoLinkedList.h"
 
-#include "CServerAIEventNode.h"
-#include "CNWSObjectActionNode.h"
-#include "CNWSCombatRoundAction.h"
-#include "CNWSClient.h"
-#include "CExoKeyTable.h"
-#include "CNWSPlayerTURD.h"
-#include "CLastUpdateObject.h"
-#include "C2DA.h"
-#include "CRes.h"
-#include "CERFString.h"
-#include "CExoString.h"
-#include "CERFKey.h"
-#include "CERFRes.h"
-#include "CResRef.h"
 template <>
 int CExoLinkedList<CServerAIEventNode>::AddAfter(CServerAIEventNode *, CExoLinkedListNode *)
 {
@@ -1143,7 +1129,7 @@ int CExoLinkedList<CNWSDialogPlayer>::AddTail(CNWSDialogPlayer *)
 }
 
 template <>
-unsigned long CExoLinkedList<CNWSDialogPlayer>::Count()
+int CExoLinkedList<CNWSDialogPlayer>::Count()
 {
 	asm("leave");
 	asm("mov $0x0823e70c, %eax");
